Created new plugin and interface for model file io
