+++ bundles/org.palladiosimulator.simexp.model.io/META-INF/MANIFEST.MF
@@ -1,0 +1,19 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Io
+Bundle-SymbolicName: org.palladiosimulator.simexp.model.io
+Bundle-Version: 1.0.0.qualifier
+Import-Package: org.eclipse.emf.cdo;version="4.20.0",
+ org.eclipse.emf.cdo.common.id;version="4.20.0",
+ org.eclipse.emf.common.util;version="2.28.0",
+ org.eclipse.emf.ecore;version="2.33.0",
+ org.eclipse.emf.ecore.resource,
+ org.palladiosimulator.envdyn.environment.dynamicmodel,
+ org.palladiosimulator.experimentautomation.experiments,
+ org.palladiosimulator.monitorrepository,
+ org.palladiosimulator.pcm.core.entity,
+ tools.mdsd.modelingfoundations.identifier
+Require-Bundle: org.eclipse.emf.ecore,
+ org.palladiosimulator.envdyn.model
+Automatic-Module-Name: org.palladiosimulator.simexp.model.io
+Bundle-RequiredExecutionEnvironment: JavaSE-17
