+++ bundles/org.palladiosimulator.simexp.model.io/src/org/palladiosimulator/simexp/model/io/GroundProbabilisticNetworkLoader.java
@@ -1,20 +1,21 @@
-package org.palladiosimulator.simexp.pcm.examples.executor;
+package org.palladiosimulator.simexp.model.io;
 
 import org.eclipse.emf.common.util.EList;
 import org.eclipse.emf.common.util.URI;
 import org.eclipse.emf.ecore.EObject;
 import org.eclipse.emf.ecore.resource.Resource;
 import org.eclipse.emf.ecore.resource.ResourceSet;
-import org.palladiosimulator.monitorrepository.MonitorRepository;
+import org.palladiosimulator.envdyn.environment.staticmodel.GroundProbabilisticNetwork;
 
-public class MonitorRepositoryLoader {
+public class GroundProbabilisticNetworkLoader implements ModelLoader<GroundProbabilisticNetwork> {
 	
-	public MonitorRepository load(ResourceSet rs, URI uri) {
+	@Override
+	public GroundProbabilisticNetwork load(ResourceSet rs, URI uri) {
 		Resource resource = rs.getResource(uri, true);
 		EList<EObject> contents = resource.getContents();
 	    if (contents.isEmpty())  {
 	        return null;
 	    }
-	    return (MonitorRepository) contents.get(0);
+	    return (GroundProbabilisticNetwork) contents.get(0);
 	}
 }
