+++ bundles/org.palladiosimulator.simexp.model.io/src/org/palladiosimulator/simexp/model/io/ModelLoader.java
@@ -1,0 +1,8 @@
+package org.palladiosimulator.simexp.model.io;
+
+import org.eclipse.emf.common.util.URI;
+import org.eclipse.emf.ecore.resource.ResourceSet;
+
+public interface ModelLoader<T> {
+    T load(ResourceSet rs, URI uri);
+}
