+++ bundles/org.palladiosimulator.simexp.model.io/src/org/palladiosimulator/simexp/model/io/MonitorRepositoryLoader.java
@@ -1,20 +1,21 @@
-package org.palladiosimulator.simexp.pcm.examples.executor;
+package org.palladiosimulator.simexp.model.io;
 
 import org.eclipse.emf.common.util.EList;
 import org.eclipse.emf.common.util.URI;
 import org.eclipse.emf.ecore.EObject;
 import org.eclipse.emf.ecore.resource.Resource;
 import org.eclipse.emf.ecore.resource.ResourceSet;
-import org.palladiosimulator.envdyn.environment.staticmodel.GroundProbabilisticNetwork;
+import org.palladiosimulator.monitorrepository.MonitorRepository;
 
-public class GroundProbabilisticNetworkLoader {
+public class MonitorRepositoryLoader implements ModelLoader<MonitorRepository> {
 	
-	public GroundProbabilisticNetwork load(ResourceSet rs, URI uri) {
+	@Override
+	public MonitorRepository load(ResourceSet rs, URI uri) {
 		Resource resource = rs.getResource(uri, true);
 		EList<EObject> contents = resource.getContents();
 	    if (contents.isEmpty())  {
 	        return null;
 	    }
-	    return (GroundProbabilisticNetwork) contents.get(0);
+	    return (MonitorRepository) contents.get(0);
 	}
 }
